rainbow pluse bug fix

+++ Project/src/LightPwmDrv.c
@@ -431,7 +431,7 @@ void LightRGBWBRCtrl(uint8_t RValue, uint8_t GValue, uint8_t BValue, uint8_t WVa
   */
   
   // Percentage to Pulse Width
-  uint16_t usPW_Scale = BRPercent * (TIM2_PWM_PULSE / 100);
+  uint32_t usPW_Scale = BRPercent * (TIM2_PWM_PULSE / 100);
   uint16_t usPW_R = RValue * WATT_R_PERCENTAGE / 100 * usPW_Scale / 255;
   uint16_t usPW_G = GValue * WATT_G_PERCENTAGE / 100 * usPW_Scale / 255;
   uint16_t usPW_B = BValue * WATT_B_PERCENTAGE / 100 * usPW_Scale / 255;
